# fix sorting of commits in diff view