<!---
Thank you for contributing to GitUI! Please fill out the template below, and remove or add any
information as you feel neccesary.
--->
Commits were sorted by time in diff view. That's wrong as some people like to commit in the future or past.
To see the fix just compare my commit to one before by selecting both and pressing SHIFT+C.

This Pull Request fixes/closes #{issue_num}.

It changes the following:
- diff view is correct now


I followed the checklist:
- ( ) I added unittests
- [x] I ran `make check` without errors
- [x] I tested the overall application
- [x] I tested my changes
- [x] I added an appropriate item to the changelog