@extrawurst 2024-01-08
@Joshix-1 the problem with this change is that it is right in the hotpath for giant repos like linux. can you benchmark the differences with and without your change?