@Joshix-1 2024-01-08
How and what can I benchmark?

Shouldn't it improve performance as commits aren't unneccessarily sorted by date in most cases?