@extrawurst 2024-02-12
check the log for this line (after the commits are all collected in the log tab):

`13:47:30 [TRACE] (4) scopetime: [scopetime/src/lib.rs:41] scopetime: 2360 ms [asyncgit::revlog::async::revlog] @asyncgit/src/revlog.rs:176` (after opening kubernetes repo)

`13:57:42 [TRACE] (4) scopetime: [scopetime/src/lib.rs:41] scopetime: 26058 ms [asyncgit::revlog::async::revlog] @asyncgit/src/revlog.rs:176` (after opening linux repo)