@Joshix-1 2024-02-12
my gitui is in `~/code/gitui` i built it with `cargo build --features=timing --release`
I freshly cloned both repos at their current master, without --filter

# Linux:

## this pr:
```
~/code/linux master $ ~/code/gitui/target/release//gitui -l ; grep asyncgit::revlog::async::revlog ~/.cache/gitui/gitui.log
Logging enabled. log written to: "/home/josh/.cache/gitui/gitui.log"
17:39:51 [TRACE] (3) scopetime: [scopetime/src/lib.rs:41] scopetime: 35821 ms [asyncgit::revlog::async::revlog] @asyncgit/src/revlog.rs:176
~/code/linux master $ rm ~/.cache/gitui/gitui.log && ~/code/gitui/target/release//gitui -l && grep asyncgit::revlog::async::revlog ~/.cache/gitui/gitui.log
Logging enabled. log written to: "/home/josh/.cache/gitui/gitui.log"
17:41:54 [TRACE] (4) scopetime: [scopetime/src/lib.rs:41] scopetime: 37788 ms [asyncgit::revlog::async::revlog] @asyncgit/src/revlog.rs:176
```

## current master
```
~/code/linux master $ rm ~/.cache/gitui/gitui.log && ~/code/gitui/target/release//gitui -l && grep asyncgit::revlog::async::revlog ~/.cache/gitui/gitui.log
Logging enabled. log written to: "/home/josh/.cache/gitui/gitui.log"
17:55:57 [TRACE] (4) scopetime: [scopetime/src/lib.rs:41] scopetime: 41685 ms [asyncgit::revlog::async::revlog] @asyncgit/src/revlog.rs:142
~/code/linux master $ rm ~/.cache/gitui/gitui.log && ~/code/gitui/target/release//gitui -l && grep asyncgit::revlog::async::revlog ~/.cache/gitui/gitui.log
Logging enabled. log written to: "/home/josh/.cache/gitui/gitui.log"
17:56:54 [TRACE] (4) scopetime: [scopetime/src/lib.rs:41] scopetime: 43782 ms [asyncgit::revlog::async::revlog] @asyncgit/src/revlog.rs:142
```

# Kubernetes

## this pr
```
~/code/kubernetes master $ rm ~/.cache/gitui/gitui.log && ~/code/gitui/target/release//gitui -l && grep asyncgit::revlog::async::revlog ~/.cache/gitui/gitui.log
Logging enabled. log written to: "/home/josh/.cache/gitui/gitui.log"
17:47:54 [TRACE] (4) scopetime: [scopetime/src/lib.rs:41] scopetime: 2382 ms [asyncgit::revlog::async::revlog] @asyncgit/src/revlog.rs:176
~/code/kubernetes master $ rm ~/.cache/gitui/gitui.log && ~/code/gitui/target/release//gitui -l && grep asyncgit::revlog::async::revlog ~/.cache/gitui/gitui.log
Logging enabled. log written to: "/home/josh/.cache/gitui/gitui.log"
17:48:38 [TRACE] (3) scopetime: [scopetime/src/lib.rs:41] scopetime: 2364 ms [asyncgit::revlog::async::revlog] @asyncgit/src/revlog.rs:176
```

## current master
```
~/code/kubernetes master $ rm ~/.cache/gitui/gitui.log && ~/code/gitui/target/release//gitui -l && grep asyncgit::revlog::async::revlog ~/.cache/gitui/gitui.log
Logging enabled. log written to: "/home/josh/.cache/gitui/gitui.log"
17:53:30 [TRACE] (4) scopetime: [scopetime/src/lib.rs:41] scopetime: 2640 ms [asyncgit::revlog::async::revlog] @asyncgit/src/revlog.rs:142
~/code/kubernetes master $ rm ~/.cache/gitui/gitui.log && ~/code/gitui/target/release//gitui -l && grep asyncgit::revlog::async::revlog ~/.cache/gitui/gitui.log
Logging enabled. log written to: "/home/josh/.cache/gitui/gitui.log"
17:53:38 [TRACE] (4) scopetime: [scopetime/src/lib.rs:41] scopetime: 2378 ms [asyncgit::revlog::async::revlog] @asyncgit/src/revlog.rs:142
```